Fix BUILD operation
The final object should be based on the current instance, what the old
code calls the "standin".

Using the state to update this object should not touch `.ext`.

As the FIXME notes, this probably still wrong as we should be updating
the memoized instance by reference, not updating a cloned version of it.
However, this is at least "more correct" and I need a testcase for that
first.